backend/models/ml_predictor.py
"""
Enhanced ML Predictor
Uses feature engineering and basic ML for better predictions
"""

import yfinance as yf
import numpy as np
import pandas as pd
from sklearn.ensemble import GradientBoostingRegressor
from datetime import datetime, timedelta
import warnings
warnings.filterwarnings('ignore')

class EnhancedPredictor:
    def __init__(self):
        self.model = None
        
    def get_historical_data(self, symbol, days=30):
        """Get historical price data"""
        ticker = yf.Ticker(symbol)
        data = ticker.history(period=f"{days}d", interval="1h")
        return data
    
    def create_features(self, df):
        """Create technical indicators and features"""
        features = pd.DataFrame()
        
        # Price-based features
        features['returns'] = df['Close'].pct_change()
        
        # Moving averages
        features['sma_5'] = df['Close'].rolling(window=5).mean()
        features['sma_10'] = df['Close'].rolling(window=10).mean()
        
        # Volatility
        features['volatility'] = df['Close'].rolling(window=10).std()
        
        # Volume
        features['volume_norm'] = df['Volume'] / df['Volume'].rolling(window=10).mean()
        
        # Momentum
        features['momentum'] = df['Close'] - df['Close'].shift(5)
        
        # Replace inf and NaN
        features = features.replace([np.inf, -np.inf], np.nan)
        features = features.ffill().fillna(0)
        
        return features
    
    def prepare_training_data(self, df, horizon_minutes=30):
        """Prepare X, y for training"""
        features = self.create_features(df)
        
        # Target: price change in next N minutes
        # For hourly data, horizon_minutes=30 means 0.5 hours ahead
        shift_periods = int(horizon_minutes / 60)  # Convert to hours
        if shift_periods < 1:
            shift_periods = 1
            
        df['target'] = df['Close'].shift(-shift_periods)
        df['target_change_pct'] = ((df['target'] - df['Close']) / df['Close']) * 100
        
        # Combine features with target
        data = pd.concat([features, df[['target_change_pct']]], axis=1)
        data = data.dropna()
        
        if len(data) < 10:
            return None, None
        
        X = data.drop('target_change_pct', axis=1)
        y = data['target_change_pct']
        
        return X, y
    
    def train(self, symbol, horizon_minutes=30):
        """Train the model on historical data"""
        try:
            # Get data
            df = self.get_historical_data(symbol)
            
            if df.empty or len(df) < 50:
                return False
            
            # Prepare training data
            X, y = self.prepare_training_data(df, horizon_minutes)
            
            if X is None or len(X) < 10:
                return False
            
            # Train model
            self.model = GradientBoostingRegressor(
                n_estimators=50,
                max_depth=3,
                learning_rate=0.1,
                random_state=42
            )
            
            self.model.fit(X, y)
            return True
            
        except Exception as e:
            print(f"Training error: {e}")
            return False
    
    def predict(self, symbol, horizon_minutes=30):
        """Make prediction for a symbol"""
        try:
            # Get recent data
            df = self.get_historical_data(symbol, days=7)
            
            if df.empty:
                return None
            
            # Create features for latest data point
            features = self.create_features(df)
            latest_features = features.iloc[-1:].values
            
            # Check for NaN
            if np.isnan(latest_features).any():
                latest_features = np.nan_to_num(latest_features, nan=0.0)
            
            # Train if model doesn't exist
            if self.model is None:
                success = self.train(symbol, horizon_minutes)
                if not success:
                    return None
            
            # Predict
            predicted_change = self.model.predict(latest_features)[0]
            
            # Get current price
            current_price = float(df['Close'].iloc[-1])
            
            # Calculate predicted price
            predicted_price = current_price * (1 + predicted_change / 100)
            
            # Calculate confidence based on model consistency
            # Simple heuristic: lower volatility = higher confidence
            recent_volatility = df['Close'].pct_change().tail(10).std() * 100
            base_confidence = 85
            confidence = max(60, min(95, base_confidence - recent_volatility * 10))
            
            # Confidence intervals (wider for higher volatility)
            uncertainty = abs(predicted_change) * 0.5 + recent_volatility
            min_price = predicted_price * (1 - uncertainty / 100)
            max_price = predicted_price * (1 + uncertainty / 100)
            
            return {
                'current_price': round(current_price, 2),
                'predicted_change_pct': round(predicted_change, 2),
                'predicted_price': round(predicted_price, 2),
                'confidence': round(confidence, 1),
                'min_price': round(min_price, 2),
                'max_price': round(max_price, 2),
                'volatility': round(recent_volatility, 2)
            }
            
        except Exception as e:
            print(f"Prediction error for {symbol}: {e}")
            return None


def demo_ml_predictions():
    """Demo the ML predictor"""
    print("=" * 70)
    print("🧠 ENHANCED ML PREDICTIONS - Demo")
    print("=" * 70)
    print()
    
    predictor = EnhancedPredictor()
    
    assets = {
        'BTC': 'BTC-USD',
        'ETH': 'ETH-USD',
        'GOLD': 'GC=F',
    }
    
    for asset_id, symbol in assets.items():
        print(f"\n📊 {asset_id}")
        print("-" * 70)
        
        # Get predictions for multiple horizons
        horizons = [
            (10, "10 min"),
            (20, "20 min"),
            (30, "30 min"),
        ]
        
        for minutes, label in horizons:
            result = predictor.predict(symbol, horizon_minutes=minutes)
            
            if result:
                arrow = "🔼" if result['predicted_change_pct'] > 0 else "🔽"
                print(f"\n   {label:8} {arrow} {result['predicted_change_pct']:+.2f}%")
                print(f"            Current:  ${result['current_price']:,.2f}")
                print(f"            Predicted: ${result['predicted_price']:,.2f}")
                print(f"            Range:     ${result['min_price']:,.2f} - ${result['max_price']:,.2f}")
                print(f"            Confidence: {result['confidence']:.0f}%")
                print(f"            Volatility: {result['volatility']:.2f}%")
            else:
                print(f"   {label:8} ❌ Prediction failed")
    
    print("\n" + "=" * 70)
    print("✅ ML Predictions Complete!")
    print("\n💡 Features used:")
    print("   • Moving averages (5, 10, 20 periods)")
    print("   • Volatility & momentum indicators")
    print("   • Volume trends")
    print("   • Price position & ROC")
    print("=" * 70)


if __name__ == "__main__":
    demo_ml_predictions()
