backend/requirements.txt
# Financial Prediction App - Backend Requirements

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
asyncpg==0.30.0

# TimescaleDB support
timescale-vector==0.0.4

# Redis
redis==5.2.0
aioredis==2.0.1

# Data Collection
yfinance==0.2.66
requests==2.32.5
python-binance==1.0.22
beautifulsoup4==4.14.2

# Environment & Config
python-dotenv==1.2.1

# Machine Learning
scikit-learn==1.6.0
lightgbm==4.5.0
xgboost==2.1.3
numpy==2.3.4
pandas==2.3.3

# NLP & Sentiment Analysis
transformers==4.48.0
torch==2.6.0
sentencepiece==0.2.0

# Data Processing
python-dateutil==2.9.0
pytz==2025.2

# HTTP & API
httpx==0.28.1
aiohttp==3.11.11

# Logging & Monitoring
python-json-logger==3.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Utilities
tenacity==9.0.0  # Retry logic
schedule==1.2.2  # Task scheduling

# Security
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
slowapi==0.1.9  # Rate limiting
